requirements.txt
fastapi==0.128.0
numpy==2.4.1
paddleocr==3.3.2
Pillow==12.1.0
scikit_learn==1.8.0
paddlepaddle
python-multipart